Finishing freetype comparing, moving compare function to another file
